Only require ref to exist when using tag/branch
Changes to the GitHub GraphQL API mean that resolving
references with `HEAD` no longer work.

However, we don't need to actually do this when
providing a commit SHA as a base,
so we can move the error check to only when
we need this information when committing to the same
branch as the base.